chore!: rename types submodule to service_info
