+++ bento_lib/__init__.py
@@ -6,6 +6,7 @@ from . import drs
 from . import events
 from . import schemas
 from . import search
+from . import service_info
 from . import workflows
 
 
@@ -15,4 +16,4 @@ config.read(os.path.join(os.path.dirname(os.path.realpath(__file__)), "package.c
 
 name = config["package"]["name"]
 __version__ = config["package"]["version"]
-__all__ = ["__version__", "auth", "drs", "events", "schemas", "search", "workflows"]
+__all__ = ["__version__", "auth", "drs", "events", "schemas", "search", "service_info", "workflows"]
